seperate scoring
