+++ experiments/src/main/java/se/de/hu_berlin/informatik/experiments/defects4j/calls/ERProducePredicates.java
@@ -28,10 +28,7 @@ import soot.toolkits.graph.UnitGraph;
 import soot.toolkits.graph.pdg.HashMutablePDG;
 import soot.toolkits.graph.pdg.PDGRegion;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.FileWriter;
-import java.io.IOException;
+import java.io.*;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
@@ -120,35 +117,15 @@ public class ERProducePredicates extends AbstractProcessor<BuggyFixedEntity<?>, 
                 buggyTestBinDir, buggyTestCP, testClassesFile,
                 rankingDir.resolve(subDirName), failingTests);
 
-
-
         // mine Signatures
         Miner miner = new Miner();
         HashMap<Signature.Identifier, Signature> signatures = miner.mine(folder);
 
+        this.writeToFile(folder, "signatures.dat", signatures);
+
         //resolve Code locations
         Output.readFromFile(folder);
         Output.writeToHumanFile(folder);
-
-        signatures.values().forEach(signature -> {
-            signature.setPredicates();
-            for (Predicate predicate : signature.predicates) {
-                signature.locations.addAll(predicate.getLocation());
-            }
-        });
-
-        Log.out(this, "getTargets %s.", folder);
-        List<CodeLocation> targets =  getTargets(buggyEntity);
-
-        Log.out(this, "Scoring %s.", folder);
-        double score = getScore(signatures, targets, buggyEntity);
-
-        Log.out(this, "Finish scoring %s.", folder);
-
-
-        //Output
-        String line = buggyEntity.getUniqueIdentifier() + ";" + score;
-        ScoringFileWriter.getInstance().write(line);
 
         System.out.println();
         signatures.forEach((identifier, signature) -> System.out.println("DS: " + identifier.DS
@@ -157,7 +134,6 @@ public class ERProducePredicates extends AbstractProcessor<BuggyFixedEntity<?>, 
                 + " ); "
                 //+ Arrays.toString(signature.allItems.stream().map(item -> item.prefixedId).toArray()));
                 + signature.locations));
-        Log.out(this, "Score for %s is %f.", folder, score);
 
         System.out.println();
         signatures.forEach((identifier, signature) ->  writeSignatures("DS: " + identifier.DS
@@ -235,150 +211,18 @@ public class ERProducePredicates extends AbstractProcessor<BuggyFixedEntity<?>, 
         return buggyEntity;
     }
 
-    private double getScore(HashMap<Signature.Identifier, Signature> signatures, List<CodeLocation> targets, BuggyFixedEntity<?> buggyEntity) {
-        List<Double> scores = new ArrayList<>();
-        if (signatures.size() == 0)
-            Log.out(this, "score with %s signatures in %s", signatures.size() , buggyEntity.getUniqueIdentifier());
-        if (targets.size() == 0)
-            Log.out(this, "score with %s targets in %s", targets.size(), buggyEntity.getUniqueIdentifier());
-        signatures.forEach((key, signature) -> {
-            signature.predicates.forEach(predicate -> {
-                predicate.getLocation().forEach(predicateLocation -> {
-                    for (CodeLocation target : targets) {
-                        if (scores.contains(0.0)) //skip calculating if we got a full hit
-                            continue;
-                        scores.add(calculateScore(predicateLocation, target, buggyEntity));
-                    }
-                });
-            });
-        });
-        Optional<Double> aDouble = scores.stream().min(Double::compareTo);
-
-        return aDouble.orElse(Double.NaN);
-    }
-
-    private double calculateScore(String predicateLocation, CodeLocation target, BuggyFixedEntity<?> buggyEntity) {
-        //Log.out(this, "calculateScore for  %s", predicateLocation);
-
-        LinkedList<CodeLocation> predicateLocations =  new LinkedList<>();
-
-        String path = predicateLocation.split(":")[0];
-        int predicateLine = Integer.parseInt(predicateLocation.split(":")[1]);
-        String[] packageAndClassPath = path.split("/");
-        String[] packagePath = Arrays.copyOf(packageAndClassPath, packageAndClassPath.length - 1);
-        String pck = String.join(".", packagePath);
-        String className = packageAndClassPath[packageAndClassPath.length - 1];
-        String classPaths = buggyEntity.getBuggyVersion().getClassPath(true) + ":" + buggyEntity.getBuggyVersion().getTestClassPath(true);
-        SootConnector sc = SootConnector.getInstance(pck, className, classPaths, false);
-        List<SootMethod> methods = sc.getAllMethods();
-        //Log.out(this, "Calculating score with %s methods", methods.size());
-        for (SootMethod sootMethod : methods) {
-            if (!sootMethod.hasActiveBody())
-                continue;
-            sootMethod.getActiveBody().getUnits().forEach(unit -> {
-                if (unit.getJavaSourceStartLineNumber() == predicateLine)
-                    predicateLocations.add(new CodeLocation(unit, className, sootMethod));
-            });
+    private void writeToFile(String outputDir, String filename, Object object){
+        try {
+            ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(outputDir + "/" + filename));
+            oos.writeObject(object);
+            oos.flush();
+            oos.close();
+        } catch (Exception ex) {
+            Log.err(Output.class, ex);
         }
-
-        //Log.out(this, "Calculating score with %s predicateLocations", predicateLocations.size());
-        if (predicateLocations.isEmpty())
-            return Double.POSITIVE_INFINITY;
-
-        //Log.out(this, MessageFormat.format("calculating Score between : {0} in Method: {1} and: {2} with method: {3} and class {4}", target.getLocationString(), target.MethodName, predicateLocation, predicateLocations.getFirst().MethodName, className));
-        for (CodeLocation goal : predicateLocations) {
-            if (target.equals(goal) || target.unit.equals(goal.unit))
-                return 0.0;
-            else if ((target.className + target.methodName).equals(goal.className + goal.methodName)) {
-                Integer unitPathDistance = this.getDistanceInUnits(sc, goal.method, goal.unit, target.unit);
-                if (unitPathDistance != null)
-                    return unitPathDistance;
-            }
-            else if (target.className.equals(goal.className)) {
-                Iterator<Edge> iteratorOnCallsOutOfMethod = sc.getIteratorOnCallsOutOfMethod(goal.method);
-                List<Double> edgeDistances = new ArrayList<>();
-                while (iteratorOnCallsOutOfMethod.hasNext()) {
-                    Edge edge = iteratorOnCallsOutOfMethod.next();
-                    if (edge.tgt() == target.method) {
-                        Integer internDistanceInGoal = this.getDistanceInUnits(sc, goal.method, goal.unit, edge.srcUnit());
-                        if (internDistanceInGoal == null)
-                            continue; //we wont get a result that is fair to compare
-                        Integer internDistanceInTarget = this.getDistanceInUnits(sc, target.method, target.unit, edge.tgt().getActiveBody().getUnits().getFirst());
-                        if (internDistanceInTarget == null)
-                            continue; //we wont get a result that is fair to compare
-                        Log.out(this, "Found a edge connection!");
-                        edgeDistances.add((double) (internDistanceInGoal + internDistanceInTarget + 10)); //TODO EdgeCost
-                    }
-                }
-                return edgeDistances.stream().min(Double::compareTo).orElse(Double.NaN);
-            }
-        }
-        return Double.MAX_VALUE;
     }
 
-    private Integer getDistanceInUnits(SootConnector sc, SootMethod method, Unit unit1, Unit unit2) {
-        UnitGraph unitGraph =  sc.getCFGForMethod(method);
-        List<Unit> unitPath = unitGraph.getExtendedBasicBlockPathBetween(unit1,unit2);
-        if (unitPath == null)
-            unitPath = unitGraph.getExtendedBasicBlockPathBetween(unit2, unit1);
-        if (unitPath != null) {
-            Log.out(this, "Found a path with length %s!", unitPath.size());
-            return unitPath.size();
-        }
-        return null;
-    }
 
-    private  List<CodeLocation> getTargets(BuggyFixedEntity<?> buggyEntity) {
-        List<CodeLocation> targets = new ArrayList<>();
-        Map<String, List<Modification>> changes = buggyEntity.getAllChanges(true, true, true, true, true, true);
-        changes.forEach((s, modifications) -> {
-            modifications.forEach(modification -> {
-                String cleanedPath = modification.getClassPath().substring(0, modification.getClassPath().length() - 5); //remove .java ending
-                String[] packageAndClassPath = cleanedPath.split("/");
-                String[] packagePath = Arrays.copyOf(packageAndClassPath, packageAndClassPath.length - 1);
-                String pck = String.join(".",packagePath);
-                String className = packageAndClassPath[packageAndClassPath.length -1];
-                String classPaths = buggyEntity.getBuggyVersion().getClassPath(true) + ":" + buggyEntity.getBuggyVersion().getTestClassPath(true);
-                SootConnector sc = SootConnector.getInstance(pck, className, classPaths);
-                List<SootMethod> methods = sc.getAllMethods();
-                //Log.out(this, "Calculating score with %s methods", methods.size());
-                for (SootMethod sootMethod : methods) {
-                    if (!sootMethod.hasActiveBody())
-                        continue;
-                    UnitPatchingChain unitPatchingChain = sootMethod.getActiveBody().getUnits();
-                    for (Unit unit : unitPatchingChain) {
-                        int start = unit.getJavaSourceStartLineNumber();
-                        int end = -1;
-                        Unit succ = unitPatchingChain.getSuccOf(unit);
-                        if (succ != null)
-                            end = succ.getJavaSourceStartLineNumber();
-
-                        List<Integer> lines = this.getLinesBetween(start, end);
-
-                        for (int possibleLine : modification.getPossibleLines()) {
-                            //Log.out(this, "possibleLine %s ", possibleLine);
-                            if (lines.contains(possibleLine))
-                                targets.add(new CodeLocation(unit, className, sootMethod));
-                        }
-                    }
-                }
-            });
-        });
-        return targets;
-    }
-
-    private List<Integer> getLinesBetween(int start, int end) {
-        List<Integer> lines = new ArrayList<>();
-        if (start == -1)
-            return lines;
-        lines.add(start);
-        if (end == -1)
-            return lines;
-        for (int i = start + 1; i < end; i++) {
-            lines.add(i);
-        }
-        return lines;
-    }
 
     private void copyFile(Entity bug, File signaturesFile, String s, String s2) {
         try {
