+++ spectra-generator/src/main/java/se/de/hu_berlin/informatik/gen/spectra/predicates/extras/Profile.java
@@ -1,9 +1,10 @@
 package se.de.hu_berlin.informatik.gen.spectra.predicates.extras;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Profile {
+public class Profile implements Serializable {
     public ArrayList<Integer> predicates;
     public boolean positiveSupport;
